adaugat DateSenzorController care se ocupa cu gestionarea datelor primite de la partea hardware si schimbat structura bazei de date

+++ src/main/java/com/sorinbratosin/licenta/POJO/HomeDataDTO.java
@@ -8,6 +8,6 @@ import java.time.LocalDateTime;
 public class HomeDataDTO {
     private String lastName;
     private LocalDateTime dataIrigare;
-    private double umiditate;
+    private int umiditate;
     private LocalDateTime dataAdaugarii;
 }
